added ManyToOne functionality and separate users to access their own …

+++ src/main/java/com/taskreminder/app/service/TaskService.java
@@ -1,5 +1,6 @@
 package com.taskreminder.app.service;
 import com.taskreminder.app.entity.Task;
+import com.taskreminder.app.entity.User;
 import com.taskreminder.app.enums.TaskPriority;
 import com.taskreminder.app.enums.TaskStatus;
 import com.taskreminder.app.repository.TaskRepository;
@@ -20,25 +21,50 @@ public class TaskService {
     @Autowired
     private TaskRepository taskRepository;
 
+    @Autowired
+    private UserService userService;
+
    public List<Task> getAllTasks(){
         return taskRepository.findAll();
    }
 
     public Task addTask(Task task){
-        return taskRepository.save(task);
+        task.setUser(userService.getCurrentUser());
+       return taskRepository.save(task);
     }
 
     public Task updateTask(Task task){
-        return taskRepository.save(task);
+        User currentUser = userService.getCurrentUser();
+
+        Task existingTask = taskRepository
+                .findById(task.getId())
+                .orElseThrow(() -> new RuntimeException("Task not found"));
+
+        if (!existingTask.getUser().equals(currentUser)) {
+            throw new RuntimeException("Unauthorized");
+        }
+
+        task.setUser(currentUser);
+       return taskRepository.save(task);
     }
 
     public void deleteTask(Integer id){
+        User currentUser = userService.getCurrentUser();
+
+        Task task = taskRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Task not found"));
+
+        if (!task.getUser().equals(currentUser)) {
+            throw new RuntimeException("Unauthorized");
+        }
+
         taskRepository.deleteById(id);
     }
 
     public Optional<Task> findById(Integer id){
-
-        return taskRepository.findById(id);
+        User currentUser = userService.getCurrentUser();
+        return taskRepository.findById(id)
+                .filter(task -> task.getUser().equals(currentUser));
     }
 
 
@@ -48,26 +74,32 @@ public Page<Task> getPagedTasks(
         TaskStatus status,
         TaskPriority priority,
         String title
-) {
-    return taskRepository.findTasks(status, priority, title, pageable);
+) { User user = userService.getCurrentUser();
+    return taskRepository.findTasks(user,status, priority, title, pageable);
 }
 
     public Page<Task> findAll(Pageable pageable) {
-       return taskRepository.findAll(pageable);
+        User user = userService.getCurrentUser();
+        return taskRepository.findByUser(user, pageable);
     }
 
     public List<Task> getTaskDueToday() {
-       LocalDate today = LocalDate.now();
-        return taskRepository.findByDueDate(today.toString())
+        User user = userService.getCurrentUser();
+        LocalDate today = LocalDate.now();
+
+        return taskRepository.findByUserAndDueDate(user, today.toString())
                 .stream()
                 .filter(task -> task.getStatus() != TaskStatus.DONE)
                 .toList();
     }
 
     public List<Task> getUpcomingTasks(int days) {
-       LocalDate today = LocalDate.now();
-       LocalDate endDate = today.plusDays(days);
-        return taskRepository.findByDueDateBetween(
+        User user = userService.getCurrentUser();
+        LocalDate today = LocalDate.now();
+        LocalDate endDate = today.plusDays(days);
+
+        return taskRepository.findByUserAndDueDateBetween(
+                        user,
                         today.plusDays(1).toString(),
                         endDate.toString()
                 ).stream()
@@ -76,23 +108,26 @@ public Page<Task> getPagedTasks(
     }
 
     public List<Task> getOverdueTasks() {
-       LocalDate today = LocalDate.now();
-        return taskRepository.findByDueDateBefore(today.toString())
+        User user = userService.getCurrentUser();
+        LocalDate today = LocalDate.now();
+
+        return taskRepository.findByUserAndDueDateBefore(user, today.toString())
                 .stream()
                 .filter(task -> task.getStatus() != TaskStatus.DONE)
                 .toList();
     }
 
     public long countAllTasks() {
-        return taskRepository.count();
+        User user = userService.getCurrentUser();
+        return taskRepository.countByUser(user);
     }
 
     public long countFilteredTasks(
             TaskStatus status,
             TaskPriority priority,
             String title
-    ) {
-        return taskRepository.countTasks(status, priority, title);
+    ) {User user = userService.getCurrentUser();
+        return taskRepository.countTasks(user, status, priority, title);
     }
 
     public long countCompletedTasks(
@@ -100,12 +135,16 @@ public Page<Task> getPagedTasks(
             TaskPriority priority,
             String title
     ) {
-        // If status filter is active and not DONE → no completed tasks
+
         if (status != null && status != TaskStatus.DONE) {
             return 0;
         }
-
-        return taskRepository.countCompletedTasks(priority, title);
+        User user = userService.getCurrentUser();
+        return taskRepository.countCompletedTasks(user, priority, title);
+    }
+    public List<Task> findAllForCurrentUser() {
+        User user = userService.getCurrentUser();
+        return taskRepository.findByUser(user);
     }
 
 }
